[RestLogger] Modify _logs array type in RestLogger class
This commit modifies the type of the `_logs` array in the `RestLogger` class. The array was originally storing objects of `IFullLog` type, but now it stores objects with properties `executionDate` and `log`, representing the execution date and the log respectively. This change allows for better organization and retrieval of log data.